Rails 7.1 Incompatibility: ActionView::Helpers::FormBuilder
Due to Rails monkey patching the Object#to_json method, it is possible for an unassuming object to call the method and trigger a <SystemStackError: stack level too deep> error.

https://github.com/rails/rails/blob/v7.1.3.2/activesupport/lib/active_support/core_ext/object/json.rb#L63

We discovered a case where this was caused by an instance of ActionView::Helpers::FormBuilder. We extended MetaRequest::Event#not_encodable? to exclude this class and prevent the recursion.